pyproject.toml
[tool.poetry]
name = "coverletter-backend"
version = "0.1.0"
description = "Backend for Write me a Cover Letter (FastAPI + LangChain)"
authors = ["Zerui Li <you@example.com>"]
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {version = "^0.30.0", extras=["standard"]}
pydantic = "^2.8.2"
SQLAlchemy = "^2.0.34"
python-dotenv = "^1.0.1"
langchain = "^0.2.16"
langchain-core = "^0.2.38"
langchain-google-genai = "^2.0.4"
pdfplumber = "^0.11.4"
python-multipart = "^0.0.9"
pillow = "^10.4.0"
pytesseract = "^0.3.13"
python-magic = "^0.4.27"
langdetect = "^1.0.9"
tenacity = "^9.0.0"
playwright = "^1.47.0"
python-docx = "^1.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

